+++ src/lib/Backend/Chatarr.ts
@@ -1,17 +1,21 @@
 import fetch from 'node-fetch';
 import parse5, {
     DefaultTreeParentNode as DTPN,
-    DefaultTreeElement as DTE,
-    DefaultTreeTextNode as DTTN
+    DefaultTreeElement as DTE
 } from 'parse5';
+import { URLSearchParams, URL } from 'url';
 
-type article = { title: string, href: string, time: string };
+type Element = DTE & DTPN;
+type article = { title: string, href: string };
 
 // cache only saves the 10 most recent articles (pinned first)
 export let cache: { c: null | article[] } = { c: null };
 let interval: NodeJS.Timeout | null = null;
 
-/** * News Aggregator. */
+/**
+ * News Aggregator. 
+ * @throws {TypeError | FetchError}
+ */
 export const chatarr = async (): Promise<article[]> => {
     const res = await fetch('https://chatarr.com/ajax/world_news_ajax/world_news_ajax.php', {
         method: 'POST',
@@ -20,34 +24,38 @@ export const chatarr = async (): Promise<article[]> => {
         },
         body: 'method=fetch_all_news_on_home_page'
     });
-
     const html = await res.text();
-    const document = parse5.parseFragment(html) as DTPN;
-    const divs = document.childNodes.filter(n => n.nodeName === 'div').slice(0, 10);
+    const document = parse5.parseFragment(html) as Element;
 
-    const articles: article[] = [];
+    const divs = (document.childNodes as Element[]).filter(n => 
+        n.nodeName === 'div' && 
+        n.attrs.some(a => a.name === 'id' && a.value === 'live_search_table')
+    ).slice(0, 10);
+    
+    const articles = [];
 
-    for(const a of divs) {
-        // this was much cleaner with regular js
-        const b = (a as DTPN).childNodes.find(e => e.nodeName === 'table');
-        const c = (b as DTPN).childNodes.find(e => e.nodeName === 'tbody');
-        const d = (c as DTPN).childNodes.find(e => e.nodeName === 'tr');
-        const e = (d as DTPN).childNodes.filter(e => e.nodeName === 'td') as (DTPN & DTE)[];
+    for(const el of divs) {
+        const table = el.childNodes.find(n => n.nodeName === 'table') as Element;
+        const tbody = table.childNodes.find(n => n.nodeName === 'tbody') as Element;
 
-        // "pinned" message is overwritten by the assignment
-        const json: article = Object.assign({}, ...e.map(f => {
-            if(f.childNodes.find(el => el.nodeName === 'a')) {
-                const a = f.childNodes.find(el => el.nodeName === 'a') as DTE;
-                return { 
-                    title: a.attrs.find(el => el.name === 'title').value.trim(),
-                    href: a.attrs.find(el => el.name === 'href').value.trim() 
-                }
-            } else {
-                return { time: (f.childNodes[0] as DTTN).value.trim() }
-            }
-        }));
+        const trs = (tbody.childNodes as Element[]).filter(n => n.nodeName === 'tr')
+            .map(n => n.childNodes)
+            .flat()
+            .filter(n => n.nodeName !== '#text');
 
-        articles.push(json);
+        const not_pinned = (trs as Element[]).find(n => n.attrs.every(a => 
+            ['id', 'class'].includes(a.name) && 
+            ['not_pinned', 'make_bold'].includes(a.value))
+        );
+
+        const a = not_pinned.childNodes.find(n => n.nodeName === 'a') as Element;
+        const href = a.attrs.find(attr => attr.name === 'href').value;
+        const title = parse5.serialize(a);
+
+        articles.push({
+            title: title.trim(),
+            href: new URLSearchParams(new URL(href).search).get('url')
+        });
     }
 
     return articles;