+++ CHANGELOG.md
@@ -324,3 +324,5 @@
 * PM2 launch command updated since it doesn't support ESM.
 * Add in ``washyourlyrics`` command.
 * Add in ``chatarr`` command.
+* Add in ``colorize`` command.
+* Add in ``forgotify`` command.